+++ integrations/github/github/core/options.py
@@ -1,3 +1,4 @@
+from datetime import datetime
 from typing import Any, List, NotRequired, Optional, Required, TypedDict
 
 from github.helpers.models import RepoSearchParams
@@ -46,7 +47,7 @@ class ListPullRequestOptions(RepositoryIdentifier):
 
     states: Required[list[str]]
     max_results: Required[int]
-    since: Required[int]
+    updated_after: Required[datetime]
 
 
 class SingleIssueOptions(RepositoryIdentifier):