+++ integrations/github/tests/github/webhook/webhook_processors/test_pull_request_webhook_processor.py
@@ -34,7 +34,7 @@ def resource_config() -> GithubPullRequestConfig:
         selector=GithubPullRequestSelector(
             query="true",
             states=["open"],
-            includeExtraFields=False,
+            api="rest",
         ),
         port=PortResourceConfig(
             entity=MappingsConfig(